--- a/inference.ipynb
+++ b/inference.ipynb
@@ -507,6 +507,56 @@
    "metadata": {},
    "outputs": [],
    "source": []
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## TaylorSeer Support for Inference Acceleration"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "# Enable TaylorSeer (https://arxiv.org/abs/2503.06923) by setting enable_taylorseer=True in inference_hyper\n",
+    "# This feature is supported for tasks that involve image output, including image generation, image generation with think, editing, and edit with think.\n",
+    "\n",
+    "# Below is an example of image generation with TaylorSeer enabled.\n",
+    "inference_hyper=dict(\n",
+    "    cfg_text_scale=4.0,\n",
+    "    cfg_img_scale=1.0,\n",
+    "    cfg_interval=[0.4, 1.0],\n",
+    "    timestep_shift=3.0,\n",
+    "    num_timesteps=50,\n",
+    "    cfg_renorm_min=0.0,\n",
+    "    cfg_renorm_type=\"global\",\n",
+    "    enable_taylorseer=True,  # Enable TaylorSeer\n",
+    ")"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "prompt = \"A female cosplayer portraying an ethereal fairy or elf, wearing a flowing dress made of delicate fabrics in soft, mystical colors like emerald green and silver. She has pointed ears, a gentle, enchanting expression, and her outfit is adorned with sparkling jewels and intricate patterns. The background is a magical forest with glowing plants, mystical creatures, and a serene atmosphere.\"\n",
+    "\n",
+    "print(prompt)\n",
+    "print('-' * 10)\n",
+    "output_dict = inferencer(text=prompt, **inference_hyper)\n",
+    "display(output_dict['image'])"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": []
   }
  ],
  "metadata": {