This PR adds the support of TaylorSeer (https://arxiv.org/abs/2503.06923) for accelerating image-output related inference tasks.

**How to enable TaylorSeer**
Simply setting `enable_taylorseer=True` in `inference_hyper` when calling `inferencer`. The default value is `False`. I have added a cell on the bottom of `inference.ipynb` to demonstrate a use case with TaylorSeer enabled.

**Supported Tasks**
TaylorSeer is a diffusion cache approach. Therefore it works for image-output related tasks, including image generation (with think) and editing (with think). It has been tested for all these tasks.

**Performance**
TaylorSeer on Bagel achieves ~2.2x speedup, without notable quality degradation. Here I provide a benchmark case on image generation, using the prompt and setup described in `inference.ipynb`.

[Original Generated Image] [Time: 36s, on 1 A100 GPU]
<img width="1024" height="1024" alt="ori_t2i" src="https://github.com/user-attachments/assets/1e5eaaff-36ae-4b63-a9a0-70449ee20421" />

[TaylorSeer Generated Image] [Time: 16s, on 1 A100 GPU]
<img width="1024" height="1024" alt="taylorseer_t2i" src="https://github.com/user-attachments/assets/d081af8b-994b-44e3-a3e4-85b2da964109" />

It is observed that TaylorSeer accelerates the generation from 36s to 16s, tested on 1 A100 GPU. The image quality remains unaffected.

Please let me know if this contributes to a nice addition!

Best,
Jiaqi Han